@NEW - Exploration Map and Assets

+++ Assets/Scripts/ScriptableScripts/EnemyAttributes.cs
@@ -20,12 +20,12 @@ public class EnemyAttributes : ScriptableObject
     public int MaxMoney { get { return _maxMoney; } }
 
     [Tooltip("ATTACK STRENGTH")]
-    [SerializeField] float _atkStrength;
-    public float AtkStrength { get { return _atkStrength; } }
+    [SerializeField] int _atkStrength;
+    public int AtkStrength { get { return _atkStrength; } }
 
     [Tooltip("ARMOR TO RESIST ATTACKS")]
-    [SerializeField] float _defenseArmor;
-    public float DefenseArmor { get { return _defenseArmor; } }
+    [SerializeField] int _defenseArmor;
+    public int DefenseArmor { get { return _defenseArmor; } }
 
     [Tooltip("MOVE SPEED")]
     [SerializeField] float _moveSpeed;
